Review

+++ src/core/call.cs
@@ -40,7 +40,7 @@ public static unsafe partial class FhCall {
         new( new FhMethodLocation(0x48E6C0, 0x34C2B0) );
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate uint d_AtelGetSaveDic();
+    public delegate ushort d_AtelGetSaveDic();
     public static FhMethodHandle<d_AtelGetSaveDic> AtelGetSaveDic =>
         new( new FhMethodLocation(0x46C3A0, 0x326B60) );
 
@@ -60,12 +60,12 @@ public static unsafe partial class FhCall {
         new( new FhMethodLocation(0x390500, 0x220BC0) );
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int d_AtelPushMember();
+    public delegate byte d_AtelPushMember();
     public static FhMethodHandle<d_AtelPushMember> AtelPushMember =>
         new( new FhMethodLocation(0x46E2A0, 0x328DA0) );
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int d_AtelPopMember();
+    public delegate byte d_AtelPopMember();
     public static FhMethodHandle<d_AtelPopMember> AtelPopMember =>
         new( new FhMethodLocation(0x46DD40, 0x3287E0) );
